compiler: Fix ValuesToList length hint
ValuesToList was calculating its list length hint with the assumption
that the passed terminator was an empty list. However, if it was a list
itself this would calculate the wrong as the length of the tail list
wouldn't be included.

Fix to only pass a length hint if the terminator is a known list
element.

+++ compiler/src/main/scala/planner/Step.scala
@@ -764,7 +764,7 @@ case class Return(returnValue : Option[TempValue]) extends Step {
   *
   * SetPairCar and SetPairCdr must be called on the new pair before it is accessed or the next GC barrier
   */
-case class InitPair(result : TempValue, listLengthOpt : Option[Int] = None) extends Step with CellConsumer {
+case class InitPair(result : TempValue, listLengthOpt : Option[Long] = None) extends Step with CellConsumer {
   val inputValues = Set[TempValue]()
   lazy val outputValues = Set(result)
